requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3

gunicorn>=21.2
dj-database-url>=2.1
psycopg2-binary>=2.9

python-dotenv>=1.0
